Add print/println functions to send commands to the device
This is the same as done in Sodaq_GSM_Modem

Furthermore do a retry (3x) of sending some commands because after
a wakeup the first reply from the device can be "invalid_param".
It is a temporary workaround and it probably should be done using
a generic send function that waits for OK and retry if the
"invalid_params" is seen.